bin/: add a pychecksyntax script
The only purpose of this file is to output a pep8-formatted message for
the first syntax error in a Python file.

That way we can see syntax errors directly in vim-ale, through a custom
error linters.

(Note: we could have used flake8 which does support this, but that would
be cheating)